Upgrade projects to Visual Studio 2015 format
I opened the existing project in Visual Studio 2015, which prompted me to
upgrade it. These are the changes that the upgrade process made.

Now that Visual Studio 2015 is provided for free for home and some
commercial use, there are hopefully few people still running the older
versions of Visual Studio such as 2005.

Note that I left the target framework set to 2.0 because some people may
still be targeting older frameworks.